Restore specs to make it work with latest Solidus
