+++ .circleci/config.yml
@@ -1,48 +1,53 @@
 version: 2.1
 
 orbs:
+  # Required for feature specs.
+  browser-tools: circleci/browser-tools@1.1
+
   # Always take the latest version of the orb, this allows us to
   # run specs against Solidus supported versions only without the need
   # to change this configuration every time a Solidus version is released
   # or goes EOL.
   solidusio_extensions: solidusio/extensions@volatile
 
-commands:
-  libvips:
-    steps:
-      - run:
-          name: Install libvips
-          command: "sudo apt-get update && sudo apt-get install -y libvips"
-
 jobs:
+  run-specs-with-sqlite:
+    executor: solidusio_extensions/sqlite
+    steps:
+      - browser-tools/install-chrome
+      - solidusio_extensions/run-tests
   run-specs-with-postgres:
     executor: solidusio_extensions/postgres
     steps:
-      - libvips
+      - browser-tools/install-chrome
       - solidusio_extensions/run-tests
-      - store_artifacts:
-          path: /tmp/capybara-screenshots
   run-specs-with-mysql:
     executor: solidusio_extensions/mysql
     steps:
-      - libvips
+      - browser-tools/install-chrome
       - solidusio_extensions/run-tests
-      - store_artifacts:
-          path: /tmp/capybara-screenshots
+  lint-code:
+    executor: solidusio_extensions/sqlite-memory
+    steps:
+      - solidusio_extensions/lint-code
 
 workflows:
   "Run specs on supported Solidus versions":
     jobs:
+      - run-specs-with-sqlite
       - run-specs-with-postgres
       - run-specs-with-mysql
-  "Weekly run specs against master":
+      - lint-code
+
+  "Weekly run specs against main":
     triggers:
       - schedule:
           cron: "0 0 * * 4" # every Thursday
           filters:
             branches:
               only:
-                - master
+                - main
     jobs:
+      - run-specs-with-sqlite
       - run-specs-with-postgres
       - run-specs-with-mysql
