+++ .github/stale.yml
@@ -1,0 +1,1 @@
+_extends: .github
