+++ CHANGELOG.md
@@ -1,0 +1,3 @@
+# Changelog
+
+See https://github.com/solidusio-contrib/solidus_klarna_payments/releases or OLD_CHANGELOG.md for older versions.
