+++ Gemfile
@@ -3,19 +3,23 @@
 source 'https://rubygems.org'
 git_source(:github) { |repo| "https://github.com/#{repo}.git" }
 
-branch = ENV.fetch('SOLIDUS_BRANCH', 'master')
-solidus_git, solidus_frontend_git = if (branch == 'master') || (branch >= 'v3.2')
-                                      %w[solidusio/solidus solidusio/solidus_frontend]
-                                    else
-                                      %w[solidusio/solidus] * 2
-                                    end
-gem 'solidus', github: solidus_git, branch: branch
-gem 'solidus_frontend', github: solidus_frontend_git, branch: branch
+branch = ENV.fetch('SOLIDUS_BRANCH', 'main')
+gem 'solidus', github: 'solidusio/solidus', branch: branch
+gem 'solidus_backend', github: 'solidusio/solidus', branch: branch
+
+# The solidus_frontend gem has been pulled out since v3.2
+if branch >= 'v3.2'
+  gem 'solidus_frontend'
+elsif branch == 'main'
+  gem 'solidus_frontend', github: 'solidusio/solidus_frontend'
+else
+  gem 'solidus_frontend', github: 'solidusio/solidus', branch: branch
+end
 
 # Provides basic authentication functionality for testing parts of your engine
 gem 'solidus_auth_devise'
 
-case ENV['DB']
+case ENV.fetch('DB', nil)
 when 'mysql'
   gem 'mysql2'
 when 'postgresql'
@@ -23,6 +27,11 @@ when 'postgresql'
 else
   gem 'sqlite3'
 end
+
+# While we still support Ruby < 3 we need to workaround a limitation in
+# the 'async' gem that relies on the latest ruby, since RubyGems doesn't
+# resolve gems based on the required ruby version.
+gem 'async', '< 3' if Gem::Version.new(RUBY_VERSION) < Gem::Version.new('3')
 
 gemspec
