+++ Rakefile
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require "bundler/gem_tasks"
 require 'solidus_dev_support/rake_tasks'
 SolidusDevSupport::RakeTasks.install
