+++ app/subscribers/solidus_klarna_payments/klarna_subscriber.rb
@@ -2,11 +2,10 @@
 
 module SolidusKlarnaPayments
   module KlarnaSubscriber
-    include ::Spree::Event::Subscriber
-    include ::SolidusSupport::LegacyEventCompat::Subscriber
+    include Omnes::Subscriber
 
-    event_action :update_klarna_shipments, event_name: :order_recalculated
-    event_action :update_klarna_customer, event_name: :order_recalculated
+    handle :order_recalculated, with: :update_klarna_shipments
+    handle :order_recalculated, with: :update_klarna_customer
 
     def update_klarna_shipments(event)
       order = event.payload[:order]
