+++ bin/rails-engine
@@ -1,6 +1,4 @@
 #!/usr/bin/env ruby
-# frozen_string_literal: true
-
 # This command will automatically be run when you run "rails" with Rails gems
 # installed from the root of your application.
