+++ lib/generators/solidus_klarna_payments/install/install_generator.rb
@@ -3,9 +3,12 @@
 module SolidusKlarnaPayments
   module Generators
     class InstallGenerator < Rails::Generators::Base
-      source_root File.expand_path('../templates', __dir__)
-
       class_option :auto_run_migrations, type: :boolean, default: false
+      source_root File.expand_path('templates', __dir__)
+
+      def self.exit_on_failure?
+        true
+      end
 
       def copy_initializer
         template 'initializer.rb', 'config/initializers/solidus_klarna_payments.rb'
@@ -16,12 +19,8 @@ module SolidusKlarnaPayments
       end
 
       def add_stylesheets
-        append_file 'vendor/assets/stylesheets/spree/frontend/all.css', "//= require spree/frontend/solidus_klarna_payments\n"
-        append_file 'vendor/assets/stylesheets/spree/backend/all.css', "//= require spree/backend/solidus_klarna_payments\n"
-      end
-
-      def add_migrations
-        run 'bin/rails railties:install:migrations FROM=solidus_klarna_payments'
+        inject_into_file 'vendor/assets/stylesheets/spree/frontend/all.css', " *= require spree/frontend/solidus_klarna_payments\n", before: %r{\*/}, verbose: true # rubocop:disable Layout/LineLength
+        inject_into_file 'vendor/assets/stylesheets/spree/backend/all.css', " *= require spree/backend/solidus_klarna_payments\n", before: %r{\*/}, verbose: true # rubocop:disable Layout/LineLength
       end
 
       def mount_engine
@@ -30,8 +29,12 @@ module SolidusKlarnaPayments
         end
       end
 
+      def add_migrations
+        run 'bin/rails railties:install:migrations FROM=solidus_klarna_payments'
+      end
+
       def run_migrations
-        run_migrations = options[:auto_run_migrations] || ['', 'y', 'Y'].include?(ask('Would you like to run the migrations now? [Y/n]')) # rubocop:disable Metrics/LineLength
+        run_migrations = options[:auto_run_migrations] || ['', 'y', 'Y'].include?(ask('Would you like to run the migrations now? [Y/n]')) # rubocop:disable Layout/LineLength
         if run_migrations
           run 'bin/rails db:migrate'
         else
