+++ lib/solidus_klarna_payments/engine.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-require 'spree/core'
-require 'solidus_klarna_payments'
+require 'solidus_core'
+require 'solidus_support'
 
 module SolidusKlarnaPayments
   class Engine < Rails::Engine
@@ -24,14 +24,6 @@ module SolidusKlarnaPayments
 
     config.to_prepare do
       ::Spree::PermittedAttributes.source_attributes << :authorization_token
-    end
-
-    initializer 'solidus_klarna_payments.pub_sub' do |app|
-      unless SolidusSupport::LegacyEventCompat.using_legacy?
-        app.reloader.to_prepare do
-          SolidusKlarnaPayments::KlarnaSubscriber.omnes_subscriber.subscribe_to(::Spree::Bus)
-        end
-      end
     end
   end
 end
