+++ spec/spec_helper.rb
@@ -6,27 +6,26 @@ ENV['RAILS_ENV'] = 'test'
 # Run Coverage report
 require 'solidus_dev_support/rspec/coverage'
 
-require File.expand_path('dummy/config/environment.rb', __dir__).tap { |file|
-  # Create the dummy app if it's still missing.
-  system 'bin/rake extension:test_app' unless File.exist? file
-}
+# Create the dummy app if it's still missing.
+dummy_env = "#{__dir__}/dummy/config/environment.rb"
+system 'bin/rake extension:test_app' unless File.exist? dummy_env
+require dummy_env
 
 # Requires factories and other useful helpers defined in spree_core.
 require 'solidus_dev_support/rspec/feature_helper'
-require 'spree/testing_support/order_walkthrough'
 
 # Requires supporting ruby files with custom matchers and macros, etc,
 # in spec/support/ and its subdirectories.
-Dir[File.join(File.dirname(__FILE__), 'support/**/*.rb')].each { |f| require f }
+Dir["#{__dir__}/support/**/*.rb"].sort.each { |f| require f }
 
-# Will load Solidus core factory first and then the ones
-# defined in `lib/solidus_klarna_payments/testing_support/factories.rb`.
-# and `lib/solidus_klarna_payments/testing_support/factories`.
+# Requires factories defined in Solidus core and this extension.
+# See: lib/solidus_klarna_payments/testing_support/factories.rb
 SolidusDevSupport::TestingSupport::Factories.load_for(SolidusKlarnaPayments::Engine)
 
 RSpec.configure do |config|
   config.infer_spec_type_from_file_location!
   config.use_transactional_fixtures = false
+
   if defined?(ActiveStorage::Current)
     config.before(:all) do
       host = 'http://www.example.com'
@@ -36,5 +35,9 @@ RSpec.configure do |config|
         ActiveStorage::Current.host = host
       end
     end
+  end
+
+  if Spree.solidus_gem_version < Gem::Version.new('2.11')
+    config.extend Spree::TestingSupport::AuthorizationHelpers::Request, type: :system
   end
 end
